sqliterepo: Trim unused code, reduce the exposure
