+++ meta2v2/meta2_server.c
@@ -121,6 +121,52 @@ _post_config(struct sqlx_service_s *ss)
 	return TRUE;
 }
 
+static GError *
+sqlx_service_resolve_peers(struct sqlx_service_s *ss,
+						   const struct sqlx_name_s *n, gboolean nocache, gchar ***result)
+{
+	EXTRA_ASSERT(ss != NULL);
+	EXTRA_ASSERT(result != NULL);
+
+	GError *err = NULL;
+	gint64 seq = 1;
+	gint retry = 1;
+
+	*result = NULL;
+
+	struct oio_url_s *u = oio_url_empty ();
+	oio_url_set(u, OIOURL_NS, ss->ns_name);
+
+	err = sqlx_name_extract(n, u, NAME_SRVTYPE_META2, &seq);
+	if (!err) {
+label_retry:
+		if (nocache)
+			hc_decache_reference_service(ss->resolver, u, NAME_SRVTYPE_META2);
+
+		gchar **peers = NULL;
+		err = hc_resolve_reference_service(
+				ss->resolver, u, NAME_SRVTYPE_META2, &peers, oio_ext_get_deadline());
+		if (err == NULL) {
+			EXTRA_ASSERT(peers != NULL);
+			*result = peers;
+			peers = NULL;
+		} else {
+			EXTRA_ASSERT(peers == NULL);
+			if (retry > 0 && err->code == CODE_RANGE_NOTFOUND) {
+				// We may have asked the wrong meta1
+				hc_decache_reference(ss->resolver, u);
+				retry --;
+				goto label_retry;
+			}
+			g_prefix_error(&err, "Peer resolution error: ");
+		}
+	}
+
+	oio_url_pclean (&u);
+	return err;
+}
+
+
 int
 main(int argc, char **argv)
 {
